--- test/testdata/test_attach_certchain.pem
@@ -1,48 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIECDCCAvCgAwIBAgIUW/XGBuFP99M7RhfzNG8/VYpC8LYwDQYJKoZIhvcNAQEL
-BQAwgYUxCzAJBgNVBAYTAklOMQwwCgYDVQQIDANERUwxDDAKBgNVBAcMA0RFTDEU
-MBIGA1UECgwLZXhhbXBsZS5jb20xETAPBgNVBAsMCHNpZ3N0b3JlMREwDwYDVQQD
-DAhzaWdzdG9yZTEeMBwGCSqGSIb3DQEJARYPZm9vQGV4YW1wbGUuY29tMB4XDTIz
-MDExNzExMjQxNFoXDTI4MDExNjExMjQxNFowgY0xCzAJBgNVBAYTAklOMQwwCgYD
-VQQIDANERUwxDDAKBgNVBAcMA0RFTDEUMBIGA1UECgwLZXhhbXBsZS5jb20xFTAT
-BgNVBAsMDHNpZ3N0b3JlLXN1YjEVMBMGA1UEAwwMc2lnc3RvcmUtc3ViMR4wHAYJ
-KoZIhvcNAQkBFg9mb29AZXhhbXBsZS5jb20wggEiMA0GCSqGSIb3DQEBAQUAA4IB
-DwAwggEKAoIBAQC44+Yn+oMxy7rempqAz95ncbTN3QkwDz1YaIvTBudyMMGAOhVO
-f3bW9ZGIXgyD3IDqxTPbmSrcaYod/6lndbVO55BikgY72zLtpN7s34RiO/P7NBDF
-AjGugXXFVNBPrNpg7cR9UCzrKV0/YCT/bQc38GRbfK3rDW3CwDITer5H4HXCa4l8
-JEcTxFRCKV70UWvYBongOoYVUgK4x7noSBjhEQ5Ixbzeekx1Geu2VxtrsxjGFNL+
-qFMoiQmmQtuocdum14b3Tn4hZQSKx4vw+PzmZUwFBIfaZWOVKlsRrRq2SDHgec0P
-HjH1bGoiJzN6oFyili7S2fqbgFYXYrrP+jtNAgMBAAGjZjBkMB0GA1UdDgQWBBQW
-B/M1a+luUKSVQONO6VOwKpwS/DAfBgNVHSMEGDAWgBRm69SULL/iPHRJGshqmBcY
-HUjz0jASBgNVHRMBAf8ECDAGAQH/AgEAMA4GA1UdDwEB/wQEAwIBhjANBgkqhkiG
-9w0BAQsFAAOCAQEAgyANtu4agUvdX2JiKn1yGM6lPi88lskwekY2iosodmqBy87H
-KALdzANQ2KSNlR8YAIvf+pxxBi1LpwUuyN4EGa+pQO+jlVcBnjgWVGop4XOf/LsM
-cBOP/tg02yQKRf6B1Poy10z+G+4Q7gFngoGzAS3Awg3Tp5k5hx74MVMUu4fCUppx
-StVqF0rCmGACzOKQZMo/KrtU2E9pnW00MQQpf+1UXoFpxDtz+dgoZ16V0YQOgwYj
-rTEVHVYWzZCATRMvkyM86AL4DjIthnRdhD0S87RMnafesn7h0eiF/OZ8MDadxfcG
-A5F1aE+z+uXWV5AH81meqRyC+gKwvUxEu7dprw==
------END CERTIFICATE-----
------BEGIN CERTIFICATE-----
-MIID7TCCAtWgAwIBAgIUD2ekQvcqyU64eTDfQ5aTYL8IA/4wDQYJKoZIhvcNAQEL
-BQAwgYUxCzAJBgNVBAYTAklOMQwwCgYDVQQIDANERUwxDDAKBgNVBAcMA0RFTDEU
-MBIGA1UECgwLZXhhbXBsZS5jb20xETAPBgNVBAsMCHNpZ3N0b3JlMREwDwYDVQQD
-DAhzaWdzdG9yZTEeMBwGCSqGSIb3DQEJARYPZm9vQGV4YW1wbGUuY29tMB4XDTIz
-MDExNzExMjE0MVoXDTI4MDExNjExMjE0MVowgYUxCzAJBgNVBAYTAklOMQwwCgYD
-VQQIDANERUwxDDAKBgNVBAcMA0RFTDEUMBIGA1UECgwLZXhhbXBsZS5jb20xETAP
-BgNVBAsMCHNpZ3N0b3JlMREwDwYDVQQDDAhzaWdzdG9yZTEeMBwGCSqGSIb3DQEJ
-ARYPZm9vQGV4YW1wbGUuY29tMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKC
-AQEAqwQjQeFHlWSIWhMfWSkBWySDdla4G8FJK6XqvD/XmrRtzNV/t133+m3y6sRs
-giXN1Ak1fh7jcfZ1WzXXBiLXi6tUJheZhvhMeqqTLodDPcFZaAQKvLgvKou+rtx1
-Kg9qPiuQ2xK/cYTpGiCNrOZf9Vz/L0SLc/L0dM/pTwdySYKrKmWNnVynzPrI4HVF
-qrEpesTVZuHdysm8bSc/TLNkOu29BqROsdicReVJCvinONbJk+CseumDnd2hxST7
-WFjARQHpiMMynRmb7eW/6jt3wT6oYvpFhOhQxLVoLuLsmxMURRDXlkzHdIdtuVyW
-/qylgQGPHUgyRyYiIGbeoTiW7wIDAQABo1MwUTAdBgNVHQ4EFgQUZuvUlCy/4jx0
-SRrIapgXGB1I89IwHwYDVR0jBBgwFoAUZuvUlCy/4jx0SRrIapgXGB1I89IwDwYD
-VR0TAQH/BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAQEABteomMmnLg49fZ6X5qVZ
-wgjNZ+nTfr36AIV2XLAlC3BIwAz0ZuvlGtkGjyGKevjFGwWf+XCOf0K44/UChxoJ
-VVJlJvYADJsO0Rleafg2FmejlS+GvXRomgHpJBQJJqlDfM7C7esScBzUXu1pS8hf
-PvvUz+gK8UHyI26BaAsBwnRH8O/ceJgjJcgGeCsQJJ+779ao+38RfnkWjkdCjIbu
-rtS3+e1zb3mJqR4Xs7rMVccsbchidh/X2ZrsgXpfXyRF6KXxMgteXqdrZt21ZC4Q
-8vl389E9ikWnpLMHouZcYv9IcFAuXcOBfz4ES+aOzRBnvmsmaQ6N3gillQpza8y4
-Zg==
------END CERTIFICATE-----
