--- test/testdata/test_attach_leafcert.pem
@@ -1,24 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIEGDCCAwCgAwIBAgIUJHDn2ZYAsfge8QUcXmItoqgQpZcwDQYJKoZIhvcNAQEL
-BQAwgY0xCzAJBgNVBAYTAklOMQwwCgYDVQQIDANERUwxDDAKBgNVBAcMA0RFTDEU
-MBIGA1UECgwLZXhhbXBsZS5jb20xFTATBgNVBAsMDHNpZ3N0b3JlLXN1YjEVMBMG
-A1UEAwwMc2lnc3RvcmUtc3ViMR4wHAYJKoZIhvcNAQkBFg9mb29AZXhhbXBsZS5j
-b20wHhcNMjMwMTE3MTEzNTI4WhcNMjgwMTE2MTEzNTI4WjCBjzELMAkGA1UEBhMC
-SU4xDDAKBgNVBAgMA0RFTDEMMAoGA1UEBwwDREVMMRQwEgYDVQQKDAtleGFtcGxl
-LmNvbTEWMBQGA1UECwwNc2lnc3RvcmUtbGVhZjEWMBQGA1UEAwwNc2lnc3RvcmUt
-bGVhZjEeMBwGCSqGSIb3DQEJARYPZm9vQGV4YW1wbGUuY29tMIIBIjANBgkqhkiG
-9w0BAQEFAAOCAQ8AMIIBCgKCAQEAsA5mQ+nqq83irr3Kg1DSZkX/KRQQoLW+cAyt
-wSnRcr6QiJy8I2snWSE6UxGtFouTOM1qEYq8yj5QD7dhf3mGgv0eQ/Etk8KMNYsw
-8IG/zsYlidrOSBdHEn3YSamZUwRajYRSGjTmxj40sigkouEPdeebeuduKynH3dF3
-m4Q2C6J754xpcXpa/5g2CkMAdmMLVeOkV3dHC2chFjS3gXnHBX4tFODS05jXim2H
-aGFYP1RILlCmN5oARkxsAGV5DiZx5lE8rRqk9dYJvQiNEnbJek/rpJiY9wc0DvJk
-Uq/dWdfuB0JvXp8DKB9IMzf1QzoPfLwiRYflMSlI7cjFr2RoyQIDAQABo2wwajAf
-BgNVHSMEGDAWgBQWB/M1a+luUKSVQONO6VOwKpwS/DAJBgNVHRMEAjAAMAsGA1Ud
-DwQEAwIE8DATBgNVHSUEDDAKBggrBgEFBQcDAzAaBgNVHREEEzARgQ9mb29AZXhh
-bXBsZS5jb20wDQYJKoZIhvcNAQELBQADggEBAEq4DmlYkFDgUSIafo1ELzwdz6X0
-0hHaPQ6xJND3FRbsDu8qGcXNPLxjUXgEI/LJoRornMwt0RWIOyRKLPu+xGJf262i
-RrZXA0NMyHat68STkI+CvAFH1v6e6eZmSMrbTDP6kdYeyxbnAQwPEawyiPjcNkDA
-11rsDxm2gvLcKzBnVbvmD76JGvNwHgzV75z8bmp+XwMkErObJI8SyNUbk0d8SeUu
-DBjarwaj1YObBvc2aC9ABwnbVxFWef5LTX9i1umr6L4aecYX2FNLl+6ZLNsOn/wO
-SvayuhPKB5inMEhGG+DQOu0EFsgcqF+g9ADD3U1ZkqSq5jEvEDm+o9ScS+w=
------END CERTIFICATE-----
