--- test/testdata/test_attach_private_key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEowIBAAKCAQEAsA5mQ+nqq83irr3Kg1DSZkX/KRQQoLW+cAytwSnRcr6QiJy8
-I2snWSE6UxGtFouTOM1qEYq8yj5QD7dhf3mGgv0eQ/Etk8KMNYsw8IG/zsYlidrO
-SBdHEn3YSamZUwRajYRSGjTmxj40sigkouEPdeebeuduKynH3dF3m4Q2C6J754xp
-cXpa/5g2CkMAdmMLVeOkV3dHC2chFjS3gXnHBX4tFODS05jXim2HaGFYP1RILlCm
-N5oARkxsAGV5DiZx5lE8rRqk9dYJvQiNEnbJek/rpJiY9wc0DvJkUq/dWdfuB0Jv
-Xp8DKB9IMzf1QzoPfLwiRYflMSlI7cjFr2RoyQIDAQABAoIBAFsAJmYGO9FwCzKN
-UWcM2KQanChMgcfACV57fzhv+B4aWDT05t5b3avZW/9iDJixrnhW3POSRcE38kQd
-Ypz0YM3GLeWqfQd6MOTgPHykZFNp4i9eVO+4cHkjaS+yJVlez7uEYLfBDmP79Vr9
-GurB8C7ed2RQyYYwA2opgmx1K0ZZawYTyHDQxXTvJGLVk/0swAr0yiKS2ExGJZI3
-n59fuEc6L7Ubf/CA41WyO6BVx03WH1PokDzgZyx1wSfhSOsvk3sld1/V7L24HLHj
-ayb/VL1z4vU/cVi3c28DZOYVLAlMVEEYSYZwRdQ732d8cYGYrJoXlII6XdWuj83n
-lbGJInECgYEA5kcTGb+rRvBIwouVhgDjzmDyzd8VsArigD1G4wOh7LYazinDgxXL
-Q2rNQsUlbrG1tqv6Jm49n8ZrbkOD66I4x7588zOxVmV+Hl+6jhA7ISpkc/tKxk+H
-9OFbfU/KnCB6cFoXR9dd5Hvdu2Q6Nimugf+xZXtVdQnO65furLQGbNUCgYEAw7jV
-CWC9w0SfMTfpjzSRTFSSH2y2io/d/koS6hcOH6XU17H+4sggoicgRPsg83LPkWkg
-vmUE5PxanpxQU8tfPJncM/jFQDzzbaC6pzaxAfExc5ZFnmrxWgxHVVXQArcH8mTl
-H+94OZCqv7nqUxWhDxia/gKHgF0RCJAR089a9iUCgYEAku+jNnjg94D3THv4yuri
-cZvXEppYol4vrRhslQqKVhrU+9wWTxym+Lu7gfF+MyXeB1OxaLhLRho8Dr2x/bcj
-eAEKhbzkWXalhQgcZ+3fbQg51O/yBJxW6mCqGknBxOHMfshdto+r79FB6QLJSipK
-mdToAxMxsl6x1A68UyMEPokCgYB98y208TpstHY1OCPm8g8mra2y50S4+kgDYiu1
-1MxT9WV4wKSjz+GgN8l+WpUj1RhaHOrjWyLBMNwoAbN3mQJafzDDiS12X0LPo4U6
-IenatBn8phF7FprmNcyZz7UiaiKFDASefn5BnSUkaEGw4dRZYJ4tI/eK1Vkmx3+R
-bHi+uQKBgFkauPtQ51M92Ah1Nc/ikcOS65aG2A3VDtoFSveeV1kGALbvV8CjSVlc
-ZM+gseUUkp8RVWvQAAURvxAzDFsYaHeMrKKToOIORPOg0Pk8HNICPJWM8echHfZ8
-BAlVZ3K6i+q3pxzk/awMR8QCSubO62B0XvKZs17+p0uU5WBU+P7F
------END RSA PRIVATE KEY-----
