--- test/testdata/test_attach_rootcert.pem
@@ -1,24 +1,0 @@
------BEGIN CERTIFICATE-----
-MIID7TCCAtWgAwIBAgIUD2ekQvcqyU64eTDfQ5aTYL8IA/4wDQYJKoZIhvcNAQEL
-BQAwgYUxCzAJBgNVBAYTAklOMQwwCgYDVQQIDANERUwxDDAKBgNVBAcMA0RFTDEU
-MBIGA1UECgwLZXhhbXBsZS5jb20xETAPBgNVBAsMCHNpZ3N0b3JlMREwDwYDVQQD
-DAhzaWdzdG9yZTEeMBwGCSqGSIb3DQEJARYPZm9vQGV4YW1wbGUuY29tMB4XDTIz
-MDExNzExMjE0MVoXDTI4MDExNjExMjE0MVowgYUxCzAJBgNVBAYTAklOMQwwCgYD
-VQQIDANERUwxDDAKBgNVBAcMA0RFTDEUMBIGA1UECgwLZXhhbXBsZS5jb20xETAP
-BgNVBAsMCHNpZ3N0b3JlMREwDwYDVQQDDAhzaWdzdG9yZTEeMBwGCSqGSIb3DQEJ
-ARYPZm9vQGV4YW1wbGUuY29tMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKC
-AQEAqwQjQeFHlWSIWhMfWSkBWySDdla4G8FJK6XqvD/XmrRtzNV/t133+m3y6sRs
-giXN1Ak1fh7jcfZ1WzXXBiLXi6tUJheZhvhMeqqTLodDPcFZaAQKvLgvKou+rtx1
-Kg9qPiuQ2xK/cYTpGiCNrOZf9Vz/L0SLc/L0dM/pTwdySYKrKmWNnVynzPrI4HVF
-qrEpesTVZuHdysm8bSc/TLNkOu29BqROsdicReVJCvinONbJk+CseumDnd2hxST7
-WFjARQHpiMMynRmb7eW/6jt3wT6oYvpFhOhQxLVoLuLsmxMURRDXlkzHdIdtuVyW
-/qylgQGPHUgyRyYiIGbeoTiW7wIDAQABo1MwUTAdBgNVHQ4EFgQUZuvUlCy/4jx0
-SRrIapgXGB1I89IwHwYDVR0jBBgwFoAUZuvUlCy/4jx0SRrIapgXGB1I89IwDwYD
-VR0TAQH/BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAQEABteomMmnLg49fZ6X5qVZ
-wgjNZ+nTfr36AIV2XLAlC3BIwAz0ZuvlGtkGjyGKevjFGwWf+XCOf0K44/UChxoJ
-VVJlJvYADJsO0Rleafg2FmejlS+GvXRomgHpJBQJJqlDfM7C7esScBzUXu1pS8hf
-PvvUz+gK8UHyI26BaAsBwnRH8O/ceJgjJcgGeCsQJJ+779ao+38RfnkWjkdCjIbu
-rtS3+e1zb3mJqR4Xs7rMVccsbchidh/X2ZrsgXpfXyRF6KXxMgteXqdrZt21ZC4Q
-8vl389E9ikWnpLMHouZcYv9IcFAuXcOBfz4ES+aOzRBnvmsmaQ6N3gillQpza8y4
-Zg==
------END CERTIFICATE-----
